working on maven build
