+++ pom.xml
@@ -1,0 +1,18 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>javapan</groupId>
+	<artifactId>javapan-parent</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+	<packaging>pom</packaging>
+
+	<name>javapan-parent</name>
+	<url>http://github.com/javapan/javapan</url>
+
+	<modules>
+		<module>javapan</module>
+		<module>example</module>
+	</modules>
+
+</project>
